lambda/requirements.txt
psycopg2-binary==2.9.9
openai==1.35.3
praw==7.7.1
beautifulsoup4==4.12.3
requests==2.31.0
googlemaps==4.10.0
loguru==0.7.2
python-dotenv==1.0.0